Image Processing: Binary Encoding and Superposition + Mfer + Black and White
In this experiment, we explore quantum image processing by resizing a selected image, my nft pfp #Mfer, to a binary form and encode it into a quantum circuit. Each pixel's value is mapped to a qubit, and quantum superposition was introduced through Hadamard gates. This image processing successfully handled up to 1024 qubits!, you can see the variable pixel amounts I used below.

Setup:
Image Selection: 
I selected an image to be processed, my mfer #nft profile picture. This image served as the starting point for our quantum encoding experiment.

Image Resizing:
I resized the image to various scales, ranging from 4x4 to 32x32 pixels (see below), to reduce the complexity and allow the use of qubits to represent each pixel.

Binary Quantization:
I applied binary quantization to the resized image by setting a threshold (eg, 128) to create a binary image where each pixel was assigned a value of either 0 or 1.
quantized_pixel_value = 
{ 1 if pixel_value ≥ 128, 2 if pixel_value < 128

Quantum Encoding:
I encoded the binary quantized image into a quantum circuit. Each pixel value was mapped to a qubit in the quantum circuit, resulting in a one to one correspondence between pixels and qubits.

For pixels with value 1, we applied an X-gate to the corresponding qubit, flipping its state to |1⟩.

For pixels with value 0, we left the corresponding qubit in its initial state.

Superposition Creation:
I applied Hadamard gates to all the qubits, creating a superposition of states representing the encoded image. 

The Hadamard gate transforms the state of each qubit as follows:
H |0⟩ = 1/sqrt(2) * ( |0⟩ + |1⟩ )
H |1⟩ = 1/sqrt(2) * ( |0⟩ - |1⟩ )

Results:
On the histogram we visualize the original image, the resized image, the quantized binary image, and the quantum circuit representing the encoded image is print (that is too large to show in these). Variable image resizes below.